feat: support sqlite diesel schema models

+++ src/models/user_provider.rs
@@ -1,8 +1,12 @@
 use chrono::NaiveDateTime;
+use diesel::{Queryable, Selectable};
 
+#[derive(Queryable, Selectable)]
+#[diesel(table_name = crate::schema::user_providers)]
+#[diesel(check_for_backend(diesel::sqlite::Sqlite))]
 pub struct UserProvider {
-    id: Option<i64>,
-    user_id: i64,
+    id: Option<i32>,
+    user_id: i32,
     provider: String,
     provider_user_id: String,
     access_token: Option<String>,
@@ -11,3 +15,4 @@ pub struct UserProvider {
     created_at: Option<NaiveDateTime>,
     updated_at: Option<NaiveDateTime>,
 }
+
